Rename Assert to AssertUDF.
PIG-3367 has introduced a new keyword "ASSERT" in version 0.12.0 and thus the same named UDF can not be used any more. I've renamed the UDF to "AssertUDF" to enable the UDF in pig 0.12.0 and provided a compatible layer to not break users that are still using Pig 0.11.0.
